chore: move the useCountdownTimer to Util
